+++ libft/srcs/ft_memmove.c
@@ -13,7 +13,7 @@
 #include <string.h>
 #include "libft.h"
 
-void			*ft_memmove(void *dst, const void *src, size_t len)
+void			*ft_memmove(void *dst, const void *src, int len)
 {
 	char		*tmp;
