+++ libft/srcs/ft_memset.c
@@ -13,9 +13,9 @@
 #include <string.h>
 #include "libft.h"
 
-void			*ft_memset(void *b, int c, size_t len)
+void			*ft_memset(void *b, int c, int len)
 {
-	size_t		i;
+	int		i;
 	char		*tmp;
 
 	i = 0;
