+++ libft/srcs/ft_strsub.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-char	*ft_strsub(char const *s, unsigned int start, size_t len)
+char	*ft_strsub(char const *s, int start, int len)
 {
-	size_t	cpt;
+	int	cpt;
 	char	*tmp;
 
 	cpt = 0;
